float to double in calculate_dk

+++ src/include/_c_sr_radial_gradient_convergence.h
@@ -4,7 +4,7 @@ double _c_calculate_dw(double distance, double tSS) {
   return pow((distance * exp((-distance * distance) / tSS)), 4);
 }
 
-float _c_calculate_dk(float Gx, float Gy, float dx, float dy, float distance) {
+double _c_calculate_dk(float Gx, float Gy, float dx, float dy, float distance) {
   float Dk = fabs(Gy * dx - Gx * dy) / sqrt(Gx * Gx + Gy * Gy);
   if (isnan(Dk)) {
     Dk = distance;
